Moved icons to the icon folder
