+++ src/SharedComponents/Icon/Icons/CloseIcon.tsx
