+++ src/SharedComponents/Icon/Icons/CopyIcon.tsx
