+++ src/SharedComponents/Icon/Icons/Dice10.tsx
