+++ src/SharedComponents/Icon/Icons/Dice12.tsx
