+++ src/SharedComponents/Icon/Icons/Dice20.tsx
