+++ src/SharedComponents/Icon/Icons/Dice4.tsx
