+++ src/SharedComponents/Icon/Icons/Dice6.tsx
