+++ src/SharedComponents/Icon/Icons/Dice8.tsx
