+++ src/SharedComponents/Icon/Icons/GearsIcon.tsx
