+++ src/SharedComponents/Icon/Icons/PencilIcon.tsx
