+++ src/SharedComponents/Icon/Icons/PlusIcon.tsx
