+++ src/SharedComponents/Icon/Icons/TrashIcon.tsx
