+++ src/SharedComponents/Icon/Icons/VeryDissatisfiedIcon.tsx
